SMLMSim

Documentation for SMLMSim.

SMLMSim.InteractionDiffusion

SMLMData.SMLD2D

SMLMData.SMLD3D

SMLMSim.CTMC

SMLMSim.Camera

SMLMSim.GenericFluor

SMLMSim.IdealCamera

SMLMSim.Images

SMLMSim.InteractionDiffusion.MoleculeFrame

SMLMSim.InteractionDiffusion.MoleculeFrame

SMLMSim.InteractionDiffusion.MoleculeHistory

SMLMSim.InteractionDiffusion.MoleculeHistory

SMLMSim.InteractionDiffusion.Monomer

SMLMSim.Line2D

SMLMSim.Molecule

SMLMSim.Nmer2D

SMLMSim.Nmer3D

SMLMSim.Pattern

SMLMSim.Pattern2D

SMLMSim.Pattern3D

SMLMSim.Point2D

SMLMSim.Point3D

SMLMSim.InteractionDiffusion.apply_boundary

SMLMSim.InteractionDiffusion.apply_boundary!

SMLMSim.InteractionDiffusion.apply_boundary!

SMLMSim.InteractionDiffusion.apply_boundary!

SMLMSim.InteractionDiffusion.calc_r

SMLMSim.InteractionDiffusion.calc_θ

SMLMSim.InteractionDiffusion.calc_ϕ

SMLMSim.InteractionDiffusion.dimerize!

SMLMSim.InteractionDiffusion.gen_image

SMLMSim.InteractionDiffusion.gen_image_stack

SMLMSim.InteractionDiffusion.gen_movie

SMLMSim.InteractionDiffusion.monomerize!

SMLMSim.InteractionDiffusion.record_positions!

SMLMSim.InteractionDiffusion.show_frame

SMLMSim.InteractionDiffusion.show_frame

SMLMSim.InteractionDiffusion.smoluchowski

SMLMSim.InteractionDiffusion.update_position!

SMLMSim.InteractionDiffusion.update_position!

SMLMSim.InteractionDiffusion.update_positions!

SMLMSim.InteractionDiffusion.update_species!

SMLMSim.getnext

SMLMSim.getstate

SMLMSim.intensitytrace

SMLMSim.kineticmodel

SMLMSim.noise

SMLMSim.noise

SMLMSim.place2D

SMLMSim.rotate!

SMLMSim.rotate!

SMLMSim.rotate!

SMLMSim.sim

SMLMSim.uniform2D

SMLMSim.uniform2D

SMLMSim.uniform2D

SMLMSim.uniform3D

SMLMData.SMLD2D — Method

function SMLD2D(cam::Camera, y_microns::Vector{T}, x_microns::Vector{T}, photons::Vector{T},
framenum::Vector{Int}, datasetnum::Vector{Int}, connectID::Vector{Int} ) where {T<:Real}

Generate SMLD2D using camera structure and physical units.

SMLMData.SMLD3D — Method

function SMLD3D(cam::Camera, y_microns::Vector{T}, x_microns::Vector{T}, z_microns::Vector{T}, photons::Vector{T},
framenum::Vector{Int}, datasetnum::Vector{Int}, connectID::Vector{Int} ) where {T<:Real}

Generate SMLD3D using camera structure and physical units.

SMLMSim.CTMC — Type

CTMC

Continous Time Markov Chain

SMLMSim.Camera — Type

Camera

SMLMSim.GenericFluor — Type

GenericFluor

Defines a fluorophore

Fields

γ: photon emission rate in Hz, Default: 1e3

q: state transision matrix. Default: q=[1.0]

SMLMSim.IdealCamera — Type

IdealCamera <: Camera

A camera with no added noise.

Fields

pixelsize

xpixels

ypixels

gain

offset

IdealCamera(;
pixelsize=0.1,
xpixels::Int=256,
ypixels::Int=256,
gain=1.0,
offset=0.0,
)

SMLMSim.Images — Type

Images

Output data type

SMLMSim.Line2D — Type

Line2D <: Pattern2D

Points with uniform random distribution between 2 endpoints.

Line2D(;λ::AbstractFloat=10.0, endpoints=[(-1.0,0.0),(1.0,0.0)])

Fields

λ: linear molecule density

'endpoints': Vector of Tuple

'n': Numbor of Points = 1

'x': X position

'y': Y position

SMLMSim.Molecule — Type

Molecule

Photophysical properties of a molecule.

This is the most general type of luminecent or scattering single molecule. Inherited types will defines the properties of a class of molecules.

SMLMSim.Nmer2D — Type

Nmer2D <: Pattern2D

N molecules symmetricaly organized around a circle with diameter d

Nmer2D(;n::Int=8, d::AbstractFloat=.1)

Fields

'n': Numbor of Points = 1

'd': Diameter

'x': X position

'y': Y position

SMLMSim.Nmer3D — Type

Nmer3D <: Pattern3D

N molecules symmetricaly organized around a circle with diameter d

Nmer3D(;n::Int=8, d::AbstractFloat=.1)

Fields

'n': Numbor of Points = 1

'd': Diameter

'x': X position

'y': Y position

'z': Z position

SMLMSim.Pattern — Type

Pattern

Abstract type for structured patterns of molecules

SMLMSim.Pattern2D — Type

Pattern2D

Abstract type for structured patterns of molecules

SMLMSim.Pattern3D — Type

Pattern3D

Abstract type for structured patterns of molecules

SMLMSim.Point2D — Type

Point2D <: Pattern2D

A single 2D point.

Point2D() = new(1, [0.0], [0.0])

Fields

'n': Numbor of Points = 1

'x': X position

'y': Y position

SMLMSim.Point3D — Type

Point3D <: Pattern3D

A single 3D point.

Point3D() = new(1, [0.0], [0.0],[0.0])

Fields

'n': Numbor of Points = 1

'x': X position

'y': Y position

'z': Z position

SMLMSim.getnext — Method

getnext(ctmc::CTMC,t::AbstractFloat)

return the time and state of next transision

SMLMSim.getstate — Method

getstate(ctmc::CTMC,t::AbstractFloat)

return the state at time t

SMLMSim.intensitytrace — Method

intensitytrace(f::GenericFluor, nframes::Int, framerate::AbstractFloat;state1=1)

Calculate an intensity trace.

SMLMSim.kineticmodel — Function

function kineticmodel(smd_true::SMLMData.SMLD,f::Molecule,nframes::Int,framerate::AbstractFloat;ndatasets::Int=1,minphotons=50.0)

generate noise-free blinking model from smd_true

SMLMSim.noise — Function

noise(smd_model::SMLMData.SMLD2D,σ_psf::AbstractFloat)

Add zero mean Gaussian noise to coordinates with σ = σ_pdf/sqrt(photons)

SMLMSim.noise — Method

noise(smd_model::SMLMData.SMLD3D,σ_psf::Vector{<:AbstractFloat})

3D data requries σ_psf = [σ_x,σ_y,σ_z]

SMLMSim.place2D — Method

place2D(p::Vector{Pattern}, xsize::Real, ysize::Real)

Place the input patterns and return SMLD2D.

SMLMSim.rotate! — Method

Rotate a Pattern in 3D by premultiplying with the rotation matrix r.

SMLMSim.rotate! — Method

Rotate a Pattern in 3D by the improper Euler angles \alpha \beta \gamma.

SMLMSim.rotate! — Method

rotate!(p::Pattern,θ::Real)

Rotate a Pattern in 2D by \theta radians.

Both molecule positions and reference positions are rotated (e.g. endpoints of a line)

SMLMSim.sim — Method

sim(;
    ρ=1.0,
    σ_PSF=.13,
    minphotons=50,
    ndatasets=10,
    nframes=1000,
    framerate=50.0, 
    pattern::Pattern=Nmer2D(),
    molecule::Molecule=GenericFluor(;q=[0 50; 1e-2 0]),
    camera::Camera=IdealCamera(),
    zrange::Vector{<:Real}=[-1.0, 1.0]
)

Generate SMLD using simulation parmeters.

SMLMSim.uniform2D — Method

uniform2D(ρ, p::Pattern2D, xsize::AbstractFloat,ysize::AbstractFloat)

Create positions of molecules from uniformly randomly placed and rotated patterns.

SMLMSim.uniform2D — Method

uniform2D(p::Vector{Pattern}, xsize::Real, ysize::Real,θ::Real)

Randomly place the input patterns with fixed rotation \theta.

SMLMSim.uniform2D — Method

uniform2D(p::Vector{Pattern2D}, xsize::Real, ysize::Real)

Randomly place and rotate the input patterns.

SMLMSim.uniform3D — Method

function uniform3D(ρ,p::Pattern, xsize::Real,ysize::Real; zrange::Vector{<:Real}=[-1.0,1.0])

Create positions of molecules from uniformly randomly placed and rotated patterns.

!!! Note ρ is 2D density. 3D density is ρ/(zrange[2]-zrange[1]).